# specs/002-core-backend-auth/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: ExpenseFlow API
  description: Core Backend API for ExpenseFlow expense management system
  version: 1.0.0
  contact:
    name: ExpenseFlow Team
servers:
  - url: https://dev.expense.ii-us.com/api
    description: Development server
  - url: https://staging.expense.ii-us.com/api
    description: Staging server
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      summary: Get current user profile
      operationId: getCurrentUser
      tags:
        - Users
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Entra ID JWT token
  schemas:
...
    UserResponse:
      type: object
      required:
        - id
        - email
        - displayName
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        department:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://httpstatuses.com/401"
            title: "Unauthorized"
            status: 401
            detail: "Missing or invalid authentication token"
...